+++ js/app.js
@@ -18,7 +18,6 @@ import FeatureViaURL from "../modules/featureViaURL/model";
 import SliderView from "../modules/snippets/slider/view";
 import SliderRangeView from "../modules/snippets/slider/range/view";
 import DropdownView from "../modules/snippets/dropdown/view";
-import ClickCounterModel from "../modules/clickCounter/model";
 import MouseHoverPopupView from "../modules/mouseHover/view";
 import QuickHelpView from "../modules/quickHelp/view";
 import WindowView from "../modules/window/view";
@@ -151,10 +150,6 @@ async function loadApp () {
     new SliderView();
     new SliderRangeView();
     new DropdownView();
-
-    if (Object.prototype.hasOwnProperty.call(Config, "clickCounter") && Object.prototype.hasOwnProperty.call(Config, "desktop") && Config.clickCounter.desktop !== "" && Object.prototype.hasOwnProperty.call(Config, "mobile") && Config.clickCounter.mobile !== "") {
-        new ClickCounterModel(Config.clickCounter.desktop, Config.clickCounter.mobile, Config.clickCounter.staticLink);
-    }
 
     if (Object.prototype.hasOwnProperty.call(Config, "mouseHover")) {
         new MouseHoverPopupView(Config.mouseHover);